moved IO function into pkg util
